RavenDB-18212 Sharding - Indexes List View - React Welcome
