+++ src/Raven.Studio/package.json
@@ -68,6 +68,8 @@
         "zip-webpack-plugin": "^4.0.1"
     },
     "dependencies": {
+        "@types/react": "^17.0.40",
+        "@types/react-dom": "^17.0.13",
         "animate.css": "^3.5.2",
         "antlr4-c3": "^1.1.16",
         "antlr4ts": "^0.5.0-alpha.4",
@@ -76,6 +78,7 @@
         "bootstrap-duration-picker": "^2.1.3",
         "bootstrap-multiselect": "^0.9.13-1",
         "bootstrap-select": "^1.13.18",
+        "classnames": "^2.3.1",
         "cronstrue": "^0.8.0",
         "diff": "^4.0.2",
         "draggabilly": "^2.4.1",
@@ -101,6 +104,8 @@
         "packery": "^2.1.2",
         "prismjs": "^1.27.0",
         "rbush": "^2.0.1",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2",
         "sortablejs": "^1.14.0",
         "style-loader": "^3.3.1",
         "toastr": "^2.1.4",
