+++ src/Raven.Studio/typescript/common/developmentHelper.ts
@@ -1,12 +1,12 @@
 ﻿
 type Feature = "Sharding";
 
-type TeamMember = "Marcin" | "Danielle" | "Kwiato";
+type TeamMember = "Marcin" | "Danielle" | "Kwiato" | "ANY";
 
 export function todo(feature: Feature, member: TeamMember, message: string = "TODO") {
     
 }
 
-export function shardingTodo(member: TeamMember, message: string = "TODO") {
+export function shardingTodo(member: TeamMember = "ANY", message: string = "TODO") {
     
 }
