+++ src/Raven.Studio/typescript/models/resources/nonShardedDatabase.ts
@@ -4,6 +4,12 @@ class nonShardedDatabase extends database {
     get root(): database {
         return this;
     }
+    
+    getLocations(): databaseLocationSpecifier[] {
+        return this.nodes().map(x => ({
+            nodeTag: x
+        }));
+    }
 }
 
 export = nonShardedDatabase;
